Extract event filters modal logic into hook
